requirements.txt
# Core Deep Learning
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0

# Video Processing
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.8

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6

# LLM Integration
anthropic==0.25.0
openai==1.12.0

# Text-to-Speech
pyttsx3==2.90
gTTS==2.3.2
soundfile==0.12.1

# AWS Integration
boto3==1.28.25
sagemaker==2.175.0
awscli==1.29.25

# Data Science
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
joblib==1.3.1
pytest==7.4.0

# Model Training
tensorboard==2.13.0
wandb==0.15.5

# Video Augmentation
albumentations==1.3.1
imgaug==0.4.0

# API Documentation
flask-swagger-ui==4.11.1
flask-restful==0.3.10

# Logging
loguru==0.7.0

# File Handling
python-magic==0.4.27